refactor(aria/tree): split directives and resolve circular dependencies
Splits up the directives across files and resolve circular dependencies in them.
