+++ src/aria/tree/index.ts
@@ -6,4 +6,6 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-export {TreeItemGroup, Tree, TreeItem} from './tree';
+export {Tree} from './tree';
+export {TreeItem} from './tree-item';
+export {TreeItemGroup} from './tree-item-group';
